pyproject.toml
[project]
name = "trends"
version = "0.1.0"
description = "Google Trends monitoring dashboard for Lifeline Australia"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytrends>=4.9.2",
    "gspread>=6.0.0",
    "pandas>=2.0.0",
    "gspread-dataframe>=4.0.0",
    "python-dotenv>=1.0.0",
]
